+++ src/ScottPlot5/ScottPlot5/Markers/OpenCircle.cs
@@ -1,9 +1,10 @@
 ﻿namespace ScottPlot.Markers;
 
-internal class None : IMarker
+internal class OpenCircle : IMarker
 {
     public void Render(SKCanvas canvas, SKPaint paint, Pixel center, float size, MarkerStyle markerStyle)
     {
-        // No rendering for none marker
+        float radius = size / 2;
+        Drawing.DrawCircle(canvas, center, radius, markerStyle.LineStyle, paint);
     }
 }